Copy and app flow updates for 1.0.2
