+++ app/Entry.js
@@ -18,7 +18,7 @@ import Onboarding1 from './views/onboarding/Onboarding1';
 import Onboarding2 from './views/onboarding/Onboarding2';
 import Onboarding3 from './views/onboarding/Onboarding3';
 import Onboarding4 from './views/onboarding/Onboarding4';
-import Onboarding5 from './views/onboarding/Onboarding5';
+import Onboarding5 from './views/onboarding/Onboarding4';
 import { SettingsScreen } from './views/Settings';
 
 const Stack = createStackNavigator();
