+++ app/locales/en.json
@@ -9,6 +9,8 @@
     "done": "Done"
   },
   "history": {
+    "coming_soon": "Coming Soon",
+    "coming_soon_subtext": "Receive alerts from your local Healthcare Authority to inform you about potential exposure to COVID-19.",
     "no_exposure": "No known",
     "possible_exposure": "Possible",
     "possible_exposure_para": "It is possible you were in contact with or close to someone who tested positive for COVID-19",
@@ -55,6 +57,7 @@
     "choose_provider_title": "Choose Health Authority",
     "commitment": "Commitment",
     "commitment_para": "Safe Paths is a global movement to develop free, open-source, privacy-by-design tools for residents, public health officials, and larger communities to flatten the curve of COVID-19, reduce fear, and prevent a surveillance-state response to the pandemic.\n\nFor more information, please visit ",
+    "covid_safe_paths": "COVID Safe Paths",
     "default_news_site_name": "Safe Paths News",
     "enter_authority_url": "Enter or paste URL",
     "event_history_subtitle": "Understand your personal exposure based on information shared by health authorities.",
@@ -95,9 +98,6 @@
     "launch_screen2_subheader": "Knowledge is power.",
     "launch_screen3_header": "If you test positive, you can choose to donate your data anonymously",
     "launch_screen3_subheader": "Which helps keep your entire community safe.",
-    "launch_screen4_header": "You're in complete control. Data is only saved on your phone.",
-    "launch_screen4_subheader": "If you test positive, you alone can choose whether to share.",
-    "launch_set_up_phone": "Set up my phone",
     "legal_page_title": "Legal",
     "loading_public_data": "loading data...",
     "location_disabled_message": "COVID Safe Paths requires location services.",
