+++ app/views/Licenses.js
@@ -1,25 +1,23 @@
+/* eslint-disable react-native/no-raw-text */
 import React, { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
   BackHandler,
   Image,
   Linking,
-  ScrollView,
   StyleSheet,
   TouchableOpacity,
   View,
 } from 'react-native';
-import { WebView } from 'react-native-webview';
 
 import foreArrow from './../assets/images/foreArrow.png';
-import licenses from './../assets/LICENSE.json';
+import fontFamily from './../constants/fonts';
 import NavigationBarWrapper from '../components/NavigationBarWrapper';
 import { Typography } from '../components/Typography';
 import Colors from '../constants/colors';
 import { Theme } from '../constants/themes';
 
-const PRIVACY_POLICY_URL =
-  'https://docs.google.com/document/d/17u0f8ni9S0D4w8RCUlMMqxAlXKJAd2oiYGP8NUwkINo/edit';
+const PRIVACY_POLICY_URL = 'http://covidsafepaths.org/privacy';
 
 export const LicensesScreen = ({ navigation }) => {
   const { t } = useTranslation();
@@ -44,49 +42,40 @@ export const LicensesScreen = ({ navigation }) => {
     };
   });
 
-  function getLicenses() {
-    let result = '<html>';
-    result +=
-      '<style>  html, body { font-size: 40px; margin: 0; padding: 0; } </style>';
-    result += '<body>';
-
-    for (let i = 0; i < licenses.terms_and_licenses.length; i++) {
-      const element = licenses.terms_and_licenses[i];
-
-      result += '<H2>' + element.name + '</H2><P>';
-      result += element.text.replace(/\n/g, '<br/>');
-      result += '<hr/>';
-    }
-    result += '</body></html>';
-
-    return result;
-  }
-
   return (
     <NavigationBarWrapper
       title={t('label.legal_page_title')}
       onBackPress={backToMain}>
-      <ScrollView contentContainerStyle={styles.contentContainer}>
-        <View style={{ flex: 4 }}>
-          <WebView
-            originWhitelist={['*']}
-            source={{
-              html: getLicenses(),
-            }}
-            style={{
-              marginTop: 15,
-              backgroundColor: Colors.INTRO_WHITE_BG,
-            }}
-          />
+      <Theme use='default'>
+        <View style={styles.legalInfo}>
+          <Typography use='headline2' style={styles.section}>
+            {t('label.covid_safe_paths')}
+          </Typography>
+
+          <View style={styles.section}>
+            <Typography use='body2'>Path Check</Typography>
+            <Typography use='body2'>58 Day Street</Typography>
+            <Typography use='body2'>Box 441621</Typography>
+            <Typography use='body2'>Somerville, MA 02144</Typography>
+          </View>
+
+          <View style={styles.section}>
+            <Typography use='body1' style={styles.hyperlink}>
+              info@pathcheck.org
+            </Typography>
+            <Typography use='body1' style={styles.hyperlink}>
+              covidsafepaths.org
+            </Typography>
+          </View>
         </View>
-      </ScrollView>
-      <Theme use='charcoal'>
+
         <TouchableOpacity
           onPress={handleTermsOfUsePressed}
           style={styles.termsInfoRow}>
           <Typography
             use='headline2'
-            onPress={() => Linking.openURL(PRIVACY_POLICY_URL)}>
+            onPress={() => Linking.openURL(PRIVACY_POLICY_URL)}
+            style={styles.privacyPolicy}>
             {t('label.privacy_policy')}
           </Typography>
           <View style={styles.arrowContainer}>
@@ -99,22 +88,35 @@ export const LicensesScreen = ({ navigation }) => {
 };
 
 const styles = StyleSheet.create({
-  contentContainer: {
-    flexDirection: 'column',
-    width: '100%',
-    backgroundColor: Colors.INTRO_WHITE_BG,
-    paddingHorizontal: 26,
-    flex: 1,
-  },
   termsInfoRow: {
     flexDirection: 'row',
     justifyContent: 'space-between',
-    backgroundColor: Colors.SILVER,
+    backgroundColor: Colors.VIOLET,
     padding: 15,
+    position: 'absolute',
+    bottom: 0,
+    width: '100%',
   },
   arrowContainer: {
     alignSelf: 'center',
     paddingRight: 20,
     paddingLeft: 20,
+  },
+  hyperlink: {
+    color: Colors.BLUE_LINK,
+    fontSize: 16,
+    lineHeight: 22.5,
+    fontFamily: fontFamily.primaryRegular,
+    textDecorationLine: 'underline',
+  },
+  legalInfo: {
+    marginTop: 10,
+    marginLeft: 20,
+  },
+  section: {
+    marginVertical: 10,
+  },
+  privacyPolicy: {
+    color: Colors.WHITE,
   },
 });
