+++ e2e/helpers/onboarding.js
@@ -2,7 +2,6 @@ import FinishSetup from '../pages/FinishSetup.po.js';
 import Onboarding1 from '../pages/Onboarding1.po.js';
 import Onboarding2 from '../pages/Onboarding2.po.js';
 import Onboarding3 from '../pages/Onboarding3.po.js';
-import Onboarding4 from '../pages/Onboarding4.po.js';
 import SignEula from '../pages/SignEula.po.js';
 
 export const navigateThroughPermissions = async languageStrings => {
@@ -17,9 +16,6 @@ export const navigateThroughPermissions = async languageStrings => {
 
   await Onboarding3.isOnScreen(languageStrings);
   await Onboarding3.tapButton(languageStrings);
-
-  await Onboarding4.isOnScreen(languageStrings);
-  await Onboarding4.tapButton(languageStrings);
 };
 
 export const navigateThroughOnboarding = async languageStrings => {
