+++ e2e/onboarding/Onboarding.spec.js
@@ -2,7 +2,6 @@ import { languages } from '../helpers/language';
 import Onboarding1 from '../pages/Onboarding1.po.js';
 import Onboarding2 from '../pages/Onboarding2.po.js';
 import Onboarding3 from '../pages/Onboarding3.po.js';
-import Onboarding4 from '../pages/Onboarding4.po.js';
 import SignEula from '../pages/SignEula.po.js';
 
 describe.each(languages)(
@@ -35,10 +34,6 @@ describe.each(languages)(
         await Onboarding3.isOnScreen(languageStrings);
         await Onboarding3.takeScreenshot();
         await Onboarding3.tapButton(languageStrings);
-
-        await Onboarding4.isOnScreen(languageStrings);
-        await Onboarding4.takeScreenshot();
-        await Onboarding4.tapButton(languageStrings);
       });
 
       afterAll(async () => {
